Refactor: Complete app overhaul and address all review feedback
This commit finalizes a major refactoring of the application, transforming it from a simple Graham analyzer into a comprehensive financial dashboard. It also incorporates all feedback from multiple code reviews, significantly improving performance, structure, error handling, and robustness.

Key Improvements:
- **Performance:** Implemented Streamlit's `@st.cache_data` on API-calling functions (`get_dados_acao` and `get_selic_atual`) to prevent redundant API calls and improve app responsiveness.
- **Modular Structure:** The monolithic `utils/app_helpers.py` has been broken down into smaller, more focused modules under a new `helpers/` directory (`display_utils.py`, `plot_utils.py`, `export_utils.py`), improving code organization and maintainability.
- **Robust Error Handling:**
  - Enhanced API error handling in `services/brapi_api.py` to catch request exceptions, log them, and return user-friendly, sanitized error messages.
  - Improved exception handling in `app.py` to be more specific and provide clearer feedback to the user.
- **Code Readability:** The main application logic in `app.py` has been extracted into a separate function, reducing nesting and improving clarity.
- **File Handling:** The PDF generation process is now more robust, and the fragile `time.sleep()` for file cleanup has been removed to prevent race conditions.
